third_party/blink/web_tests/platform/win/external/wpt/editing/run/justifyfull_4001-last-expected.txt

This is a testharness.js-based test.
Found 135 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandState("justifyfull") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p></div><p align=\\"justify\\">bar</p><p>extra</p>" but got "<p style=\\"text-align:justify\\">foo</p><p align=\\"justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p></div><p align=\\"justify\\">bar</p><p>extra</p>" but got "<p style=\\"text-align:justify\\">foo</p><p align=\\"justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p>[foo]<p align=justify>bar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p align=\\"justify\\">foo</p><div style=\\"text-align:justify\\"><p>bar</p></div><p align=\\"justify\\">baz</p><p>extra</p>" but got "<p align=\\"justify\\">foo</p><p style=\\"text-align:justify\\">bar</p><p align=\\"justify\\">baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p align=\\"justify\\">foo</p><div style=\\"text-align:justify\\"><p>bar</p></div><p align=\\"justify\\">baz</p><p>extra</p>" but got "<p align=\\"justify\\">foo</p><p style=\\"text-align:justify\\">bar</p><p align=\\"justify\\">baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p align=\\"justify\\">foo</p><div style=\\"text-align:justify\\"><p>bar</p></div><p align=\\"justify\\">baz</p><p>extra</p>" but got "<p align=\\"justify\\">foo</p><p style=\\"text-align:justify\\">bar</p><p align=\\"justify\\">baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p align=\\"justify\\">foo</p><div style=\\"text-align:justify\\"><p>bar</p></div><p align=\\"justify\\">baz</p><p>extra</p>" but got "<p align=\\"justify\\">foo</p><p style=\\"text-align:justify\\">bar</p><p align=\\"justify\\">baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<p align=justify>foo<p>[bar]<p align=justify>baz<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div align=\\"justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div align=\\"justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div align=\\"justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div align=\\"justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>[foo</div>bar]<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<br>bar</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo</div><p style=\\"text-align:justify\\">bar</p><p>extra</p>"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify>fo[o</div>b]ar<p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><p>extra</p>" but got "<span style=\\"text-align:justify\\">foo</span><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandState("justifyfull") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><p>extra</p>" but got "<span style=\\"text-align:justify\\">foo</span><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandState("justifyfull") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>[foo]</span><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><p>extra</p>" but got "<span style=\\"text-align:justify\\">foo</span><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandState("justifyfull") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><p>extra</p>" but got "<span style=\\"text-align:justify\\">foo</span><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandState("justifyfull") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<span style=text-align:justify>f[o]o</span><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div><div style=\\"text-align:justify\\">baz</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo<div style=\\"text-align:justify\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div><div style=\\"text-align:justify\\">baz</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo<div style=\\"text-align:justify\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div><div style=\\"text-align:justify\\">baz</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo<div style=\\"text-align:justify\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo</div><div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div><div style=\\"text-align:justify\\">baz</div><p>extra</p>" but got "<div style=\\"text-align:justify\\">foo<div style=\\"text-align:justify\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p></div><p>extra</p>" but got "<div align=\\"nonsense\\"><p style=\\"text-align:justify\\">foo</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p></div><p>extra</p>" but got "<div align=\\"nonsense\\"><p style=\\"text-align:justify\\">foo</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["justifyfull",""]] "<div align=nonsense><p>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:inherit><p>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p></div><p>extra</p>" but got "<div style=\\"text-align:inherit\\"><p style=\\"text-align:justify\\">foo</p></div><p>extra</p>"
[FAIL] [["justifyfull",""]] "<div style=text-align:inherit><p>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:inherit><p>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><quasit><p>foo</p></quasit></div><p>extra</p>" but got "<quasit align=\\"center\\"><p style=\\"text-align:justify\\">foo</p></quasit><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><quasit><p>foo</p></quasit></div><p>extra</p>" but got "<quasit align=\\"center\\"><p style=\\"text-align:justify\\">foo</p></quasit><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><quasit><p>foo</p></quasit></div><p>extra</p>" but got "<quasit align=\\"center\\"><p style=\\"text-align:justify\\">foo</p></quasit><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><quasit><p>foo</p></quasit></div><p>extra</p>" but got "<quasit align=\\"center\\"><p style=\\"text-align:justify\\">foo</p></quasit><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<quasit align=center><p>[foo]</p></quasit><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "right" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><div dir=\\"rtl\\">foo</div></div><p>extra</p>" but got "<div dir=\\"rtl\\" style=\\"text-align:justify\\">foo</div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "right" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><div dir=\\"rtl\\">foo</div></div><p>extra</p>" but got "<div dir=\\"rtl\\" style=\\"text-align:justify\\">foo</div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "right" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:start>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><div dir=\\"rtl\\">foo</div></div><p>extra</p>" but got "<div dir=\\"rtl\\" style=\\"text-align:justify\\">foo</div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><div dir=\\"rtl\\">foo</div></div><p>extra</p>" but got "<div dir=\\"rtl\\" style=\\"text-align:justify\\">foo</div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div dir=rtl style=text-align:end>[foo]</div><p>extra" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p></div>" but got "<div style=\\"text-align:justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p>"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"justify\\"><p>foo</p> <p>bar</p></div>" but got "<div align=\\"justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p>"
[FAIL] [["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div style=text-align:justify><p>bar</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p></div>" but got "<p style=\\"text-align:justify\\">foo</p> <div style=\\"text-align:justify\\"><p>bar</p></div>"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div style=text-align:justify><p>bar</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div style=text-align:justify><p>bar</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div align=justify><p>bar</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"justify\\"><p>foo</p> <p>bar</p></div>" but got "<p style=\\"text-align:justify\\">foo</p> <div align=\\"justify\\"><p>bar</p></div>"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div align=justify><p>bar</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["justifyfull",""]] "<p>[foo]</p> <div align=justify><p>bar</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div style=\\"text-align:justify\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div style=\\"text-align:justify\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div style=\\"text-align:justify\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div style=\\"text-align:justify\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyfull",""]] "<div style=text-align:justify><p>foo</div> <p>[bar]</p> <div style=text-align:justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div align=\\"justify\\"><p>baz</p></div>"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"justify\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"justify\\"><p>foo</p></div> <p style=\\"text-align:justify\\">bar</p> <div align=\\"justify\\"><p>baz</p></div>"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["justifyfull",""]] "<div align=justify><p>foo</div> <p>[bar]</p> <div align=justify><p>baz</div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
Harness: the test ran to completion.
